I read articles about design and I will definitely make it beautiful

+++ app/src/main/java/com/example/weather/data/utils/RiseSet.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class RiseSet(
     @SerialName("sunrise")
-    val sunrise:Int? = null,
+    val sunrise:Long? = null,
     @SerialName("sunset")
-    val sunset:Int? = null
+    val sunset:Long? = null
 )
